migrated the scraper to use the database to persist and manage article urls

+++ PaperScraper.rb
@@ -5,19 +5,19 @@
 #A script to harvest comments from the Daily Mail and Guardian websites and write them to a database, keeping the two sources equally represented
 #Version 0.1
 
-#never scrapes the same URL twice while logs exist
-
 $: << File.expand_path(File.dirname(__FILE__))
 $script_path = File.expand_path(File.dirname(__FILE__))
 
 require 'open-uri'
-require 'hpricot'
-require 'mechanize'
 require 'rss/2.0'
 require 'iconv'
-require 'logger'
+require 'ostruct'
 
+require 'hpricot'
+require 'mechanize'
+require 'logger'
 require 'active_record'
+
 environment = ENV['ENVIRONMENT'] || 'development'
 dbconf = YAML::load(File.open('config/databases.yml'))[environment]
 ActiveRecord::Base.establish_connection(dbconf) 
@@ -42,7 +42,7 @@ class Comment < ActiveRecord::Base
     end
     
     def keep_only_latest_comments
-      delete_all(["created_at < ?", cutoff_timestamp])
+      delete_all(["created_at < ?", cutoff_timestamp]) if self.count > MAXIMUM_NUMBER_OF_COMMENTS
     end
 
     def cutoff_timestamp
@@ -66,14 +66,39 @@ class Comment < ActiveRecord::Base
   end
 end
 
-PAPERS = ['Mail', 'Guardian']
+class Article < ActiveRecord::Base
+  validates_uniqueness_of :url
+end
+
+class Paper < OpenStruct
+  def replenish_article_urls
+    latest_article_urls().each do |url|
+      article = Article.create(:paper => name, :url => url)
+      article.save if article.valid?
+    end
+  end
+  
+  def time_to_replenish?
+    Article.where(:paper => name).count < 30
+  end
+  
+  protected
+  def latest_article_urls
+    content = open(articles_rss_url).read
+    feed = RSS::Parser.parse(content, false)
+    urls = feed.items.reverse.collect(&:link)
+    urls.take(10)
+  end
+end
+
+PAPERS = [
+  { :name => 'Mail',     :articles_rss_url => 'http://www.dailymail.co.uk/news/headlines/index.rss' },
+  { :name => 'Guardian', :articles_rss_url => 'http://feeds.guardian.co.uk/theguardian/commentisfree/rss' }
+].collect {|map| Paper.new(map)}
 
 class Commissioner 
-   
   def commission_feeds
-    rssdruid = RSSDruid.new
-    rssdruid.maintain_feeds
-    return nil
+    RSSDruid.new.maintain_feeds
   end
    
   def balance_data
@@ -83,8 +108,7 @@ class Commissioner
   end
     
   def run_scrape
-    scraperdruid = ScraperDruid.new(@prescription)
-    scraperdruid.invoke_scraper
+    ScraperDruid.new(@prescription).invoke_scraper
   end
 
   def complete_cycle
@@ -104,112 +128,79 @@ class Commissioner
   end   
 end   
 
-#add a string function to escape single quotes
-class String
-  def escape_single_quotes
-    self.gsub(/'/, "\\\\'")
-  end
-end
-
 class Scraper
-
-  attr_reader :output
-  attr_reader :url
-  attr_reader :paper
+  def initialize(paper)
+    @paper = paper
+  end
   
-  def scrape
-    comments = download_comments
-    to_array(comments)
-    begin
-    write_data
-    rescue Exception
-    puts "#{@paper} article has no comments."
+  def scrape(article_url)
+    comments = download_comments_from(article_url)
+    if comments.empty?
+      puts "#{@paper} article has no comments."
+    else
+      persist(comments, article_url)
     end
   end
-
+  
+  protected
+  def persist(comments, article_url)
+    comments_to_save = comments.take(20)
+    comments_to_save.each do |comment| 
+      comment = Comment.create(:comment => comment, :url => article_url, :paper => @paper)
+      comment.save if comment.valid?
+    end
+    puts "Number of #{@paper} comments inserted: #{comments_to_save.size}"
+  end
 end 
 
 class DailyMailScraper < Scraper
-  
-  def initialize ( url )
-    @a = Mechanize.new
-    @url = url
-    @paper = "Daily Mail"
+  def initialize
+    super("Daily Mail")
   end
 
   protected
- 
-  def download_comments 
-    article_id = @url.slice(/article-\d{5,7}/)
-    article_id = article_id.slice(/\d{5,7}/)
-    article_path = @url.split('http://www.dailymail.co.uk')[1].to_s
-    comments_output = @a.post('http://www.dailymail.co.uk/dwr/call/plaincall/AjaxReaderComments.paginateReaderComments.dwr',
-      {'callCount'=>'1',
-      'page'=>"#{article_path}",
-      'httpSessionId'=>'',
-      'scriptSessionId'=>'',
-      'c0-scriptName'=>'AjaxReaderComments',
-      'c0-methodName'=>'paginateReaderComments',
-      'c0-id'=>'0',
-      'c0-param0'=>'string:' + article_id,
-      'c0-param1'=>'number:1',
-      'c0-param2'=>'number:100',
-      'c0-param3'=>'string:newest',
-      'batchId'=>'0' } 
-      ) 
-  end 
-  
-  def to_array (comments)
-    chopped = comments.body.split(/yourComments="(.*)"/)
-    @output = chopped[1...-1].values_at(* chopped.each_index.select {|i| i.even?})
-  end
-
-  def write_data
-    i = 0
-    @output[1..20].slice(1..-1).each do |comment| 
-      comment.gsub!(/^u[0-9][0-9]/) { |match| "\\" + match }
-      i += 1
-      Comment.create(:comment => comment, :url => url, :paper => paper).save!
+  def download_comments_from(article_url)
+    _, article_path, article_id = article_url.match(/www.dailymail.co.uk(\/.*article-(\d+).*)$/).to_a
+    5.times do
+      response = Mechanize.new.post(
+        'http://www.dailymail.co.uk/dwr/call/plaincall/AjaxReaderComments.paginateReaderComments.dwr',
+        {'callCount'=>'1',
+        'page'=>"#{article_path}",
+        'httpSessionId'=>'',
+        'scriptSessionId'=>'',
+        'c0-scriptName'=>'AjaxReaderComments',
+        'c0-methodName'=>'paginateReaderComments',
+        'c0-id'=>'0',
+        'c0-param0'=>'string:' + article_id,
+        'c0-param1'=>'number:1',
+        'c0-param2'=>'number:100',
+        'c0-param3'=>'string:newest',
+        'batchId'=>'0' } 
+      )
+      comments = response.body.scan(/yourComments="(.*)"/)
+      return comments unless comments.empty?
     end
-    puts "Number of Daily Mail comments inserted: #{i}"
-  end 
-  
+    return []
+  end
 end
 
 class GuardianScraper < Scraper
-  def initialize ( url )
-    @url = url
-    @scrape = open(url)
-    @paper = "Guardian"
+  def initialize
+    super("Guardian")
   end
 
   protected
-
-  def download_comments
-    @scrape.rewind
-    comments = Hpricot(Iconv.conv('utf-8', @scrape.charset, @scrape.readlines.join("\n"))).search("div[@class='comment-body']")
-  end
-  
-  def to_array (comments)
-    @output = comments.collect do |comment|
-      comment.inner_text[4...-4]
-      end
-  end
-
-  def write_data
-    i = 0
-    self.output[1..20].slice(1..-1).each do |comment| 
-      i += 1
-      Comment.create(:comment => comment.escape_single_quotes.strip, :url => url, :paper => paper).save!
+  def download_comments_from(article_url)
+    comment_nodes = []
+    scrape = open(article_url) do |f|
+      page_markup = Iconv.conv('utf-8', f.charset, f.read)
+      comment_nodes = Hpricot(page_markup).search("div[@class='comment-body']")
     end
-    puts "Number of Guardian comments inserted: #{i}"
-  end
+    comment_nodes.map(&:inner_text).map(&:strip)
+  end  
 end
 
-#database functions and maintenance
-
 class DataDruid
-
   def initialize
     @mail_comment_count = Comment.mail.count
     @guardian_comment_count = Comment.guardian.count
@@ -222,138 +213,46 @@ class DataDruid
   
   def prescribe_comments
     if @guardian_comment_count > (@mail_comment_count + 20)
-        prescription = ["Mail"]
+      prescription = ["Mail"]
     elsif @mail_comment_count > (@guardian_comment_count + 20)
-        prescription = ["Guardian"]
+      prescription = ["Guardian"]
     else 
-        prescription = ["Mail", "Guardian"]
+      prescription = ["Mail", "Guardian"]
     end
-  end
- 
-  private
-end
-
-class ArticleUrls
-  def get_urls
-    @urls ||= []
-    content = ""
-    open(@source) do |s| content = s.read end
-    @feed = RSS::Parser.parse(content, false)
-    @feed.items.reverse.each { |each| @urls << each.link }
-    @urls.slice(0..11)
-    end
-end
-
-class MailUrls < ArticleUrls
-  def initialize
-    @source = 'http://www.dailymail.co.uk/news/headlines/index.rss'
-    @urls = [:source => 'Daily Mail']
-  end
-end
-
-class GuardianUrls < ArticleUrls
-  def initialize
-    @source = 'http://feeds.guardian.co.uk/theguardian/commentisfree/rss'
-    @urls = [:source => 'Guardian']
   end
 end
 
 class RSSDruid
-
   def maintain_feeds
     PAPERS.each do |paper|
-      if time_to_replenish?(paper)
-        urlset = eval( paper + "Urls.new.get_urls")
-        urlset[1..10].each do |url|
-        File.open( $script_path + "/" + paper,'a') do |f|
-          f.puts url end
-        end # urlset iterator
-      end #time to replenish if
-      remove_duplicates(paper)        
-    end #papers iterator
+      paper.replenish_article_urls if paper.time_to_replenish?
+    end
   end
-
-  private
-
-  def time_to_replenish? ( file )
-    begin
-      if File.readlines($script_path + "/" + file).size < 30
-          true
-        else false
-        end
-      rescue Exception
-      FileUtils.touch($script_path + "/" + file)
-      retry
-    end
-  end            
-
-  def remove_duplicates(paper)
-    cleaned_file_contents = File.readlines($script_path + "/" + paper).uniq   
-    File.new($script_path + "/" + paper, "w").close
-    cleaned_file_contents.each do |url|
-      File.open($script_path + "/" + paper, "a") do |f| f.puts url end
-    end
-  end    
-
 end
 
 class ScraperDruid
-
   def initialize(prescription)
     @prescription = prescription
   end
 
   def invoke_scraper
     if @prescription.include? "Mail"
-      DailyMailScraper.new(FileClipper.new('Mail', 'log/story.log').get_line).scrape end
+      article = Article.where(:consumed => false, :paper => "Mail").first
+      unless article.nil?
+        article.consumed = true
+        article.save
+        DailyMailScraper.new.scrape(article.url)
+      end
+    end
     if @prescription.include? "Guardian"
-      GuardianScraper.new(FileClipper.new('Guardian', 'log/story.log').get_line).scrape end
-  end
-
-end 
-
-class FileClipper
-
-  def initialize(queuefile, logfile)
-    @queuefile = $script_path + "/" + queuefile
-    @logfile = $script_path + "/" + logfile
-  end
-
-  def get_line
-    line = next_line
-    while in_logfile?(line)
-      line = next_line
-    end
-    log_line(line)
-    line
-  end
-
-  private
-
-  def next_line
-    queue = []
-    File.open(@queuefile, "r").each_line { |line| queue << line }
-    return "No lines in queuefile" if queue.empty?
-    File.open(@queuefile, "w").truncate(0)  
-    selected_line = queue.shift
-    File.open(@queuefile, "a") { |f| f.puts (queue.join) }  
-    selected_line
-  end
-
-  def in_logfile?(string_to_check)
-    match = File.new(@logfile, "r").detect { |log_entry| log_entry == string_to_check }
-    if match == nil 
-      return false
-    else
-      return true
+      article = Article.where(:consumed => false, :paper => "Guardian").first
+      unless article.nil?
+        article.consumed = true
+        article.save
+        GuardianScraper.new.scrape(article.url)
+      end
     end
   end
-
-  def log_line(line)
-    log = File.new(@logfile, "a")
-    log.write(line + "\n")
-  end 
-
 end
     
 if ARGV[0] == "--maintain"
